fitbite-backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
Flask-CORS==5.0.0
Werkzeug==3.1.3
python-dotenv==1.0.1
gunicorn==23.0.0
PyJWT==2.9.0
requests==2.32.3
